Split yaml serialization
Serialization is now done with the moveit_serialisation package except
for MoveIt Benchmark Suite objects which is implemented in
serialization.h/cpp files.
