+++ benchmark_suite/package.xml
@@ -9,20 +9,18 @@
 	<buildtool_depend>catkin</buildtool_depend>
 
 	<depend>boost</depend>
-	<depend>tf2_ros</depend>
-	<depend>yaml-cpp</depend>
-	<depend>geometric_shapes</depend>
-	<depend>geometry_msgs</depend>
 	<depend>eigen_conversions</depend>
-	<depend>roscpp</depend>
+	<depend>moveit_benchmark_suite_resources</depend>
 	<depend>moveit_core</depend>
 	<depend>moveit_ros_planning</depend>
 	<depend>moveit_ros_planning_interface</depend>
-	<depend>visualization_msgs</depend>
+	<depend>moveit_serialization</depend>
 	<depend>moveit_visual_tools</depend>
-	<depend>urdf_to_scene</depend>
-	<depend>moveit_benchmark_suite_resources</depend>
+	<depend>roscpp</depend>
 	<depend>rosparam_shortcuts</depend>
+	<depend>tf2_ros</depend>
+	<depend>urdf_to_scene</depend>
+	<depend>visualization_msgs</depend>
 
 	<export>
 		<moveit_benchmark_suite benchmarks="${prefix}/benchmarks"/>
