+++ benchmark_suite/src/io/gnuplot.cpp
@@ -2,7 +2,7 @@
 
 #include <moveit_benchmark_suite/io.h>
 #include <moveit_benchmark_suite/io/gnuplot.h>
-#include <moveit_benchmark_suite/yaml.h>
+#include <moveit_serialization/yaml-cpp/yaml.h>
 #include <moveit_benchmark_suite/token.h>
 
 #include <cmath>
