+++ benchmark_suite/src/plot_dataset.cpp
@@ -2,7 +2,7 @@
 #include <ros/ros.h>
 #include <moveit_benchmark_suite/io/gnuplot.h>
 #include <moveit_benchmark_suite/config/gnuplot_config.h>
-#include <moveit_benchmark_suite/yaml.h>
+#include <moveit_serialization/yaml-cpp/yaml.h>
 #include <moveit_benchmark_suite/benchmark.h>
 
 using namespace moveit_benchmark_suite;
